Get tidy model output of exposure(s) on outcome(s)

To easily get tidy model output (from linear or logistic GLM, or CoxPH) for a categorical or continuous exposure, including sample size (and N cases if logistic), outcome, and model info. Idea is to make quick PheWAS trivially easy.

For all exposures, it gets the N. For categorical exposures, the N is split by group, and a row is included for the reference category

Can provide multiple exposures and/or outcomes

Usage

get_assoc(
  d,
  x,
  y,
  z,
  model = "lm",
  af = FALSE,
  note = "",
  get_fit = FALSE,
  extreme_ps = FALSE,
  scale_x = FALSE,
  scale_y = FALSE,
  inv_norm_x = FALSE,
  inv_norm_y = FALSE,
  subset_d = TRUE,
  progress = TRUE,
  verbose = FALSE,
  ...
)

Arguments

d

A data.frame or tibble. The data.

x

A string or vector of strings. The exposure variable name(s), found in `d`. (character)

y

A string. The outcome variable name, found in `d` -- if model is 'coxph' then paste the survival object here e.g., 'Surv(time, event)' where `time` and `event` are variables in `d`. (character)

z

A string. The covariate formula (e.g., " + age + sex"), found in `d`. Default is no covariates. default="" (character)

model

A string. The type of model to perform. Can be "lm", "logistic" or "coxph". default="lm" (character)

af

Logical. Is `x` categorical? I.e., include in formula as `as.factor(x)`. default=FALSE

note

A string. If you want to include a note like "All", "Males", "C282Y homozygotes" to describe the model or sample. default="" (character)

get_fit

Logical. Default is FALSE. Get model fit? (for each model: lm=R2, logistic=McFadden's pseudo-R2, coxph=Harrell's c-statistic). default=FALSE

extreme_ps

Logical. Default is FALSE. If p==0 then return "extreme p-values" as strings. default=FALSE

scale_x

Logical. Default is FALSE. Apply scale() function to exposure? default=FALSE

scale_y

Logical. Default is FALSE. Apply scale() function to outcome? default=FALSE

inv_norm_x

Logical. Apply inv_norm() function to exposure? default=FALSE

inv_norm_y

Logical. Apply inv_norm() function to outcome? default=FALSE

subset_d

Logical. Subset `d` before passing to analysis (much quicker when multiple exposures/outcomes used). default=TRUE

progress

Logical. Show progress bar from purrr `map()` function (useful when multiple exposures/outcomes provided). default=TRUE

verbose

Logical. Be verbose, default=FALSE

...

Other `tidy_ci()` options

Value

Returns a tibble - summary statistics from a model

Author

Luke Pilling

Examples

# for one outcome, equivalent to `tidy_ci(glm(weight ~ height +age+sex, d=ukb))` - with added `n`
get_assoc(x="height", y="weight", z="+age+sex", d=ukb)
#> Error in eval(expr, envir, enclos): object 'ukb' not found

# categorical exposure, binary outcome, and stratified analysis (with note)
#  - note that data can be passed using the pipe if desired
d=ukb |> filter(sex==1) |>
  get_assoc(x="smoking_status", y="chd", z="+age", model="logistic", af=TRUE, note="Males only")
#> Error in eval(expr, envir, enclos): object 'ukb' not found

# multiple exposures and/or outcomes - get pseudo R^2
x_vars = c("bmi","ldl","sbp_0_avg")
y_vars = c("chd","t2d")
res = get_assoc(x=x_vars, y=y_vars, z="+age+sex", d=ukb, model="logistic", get_fit=TRUE)
#> Error in eval(expr, envir, enclos): object 'ukb' not found
res
#> Error in eval(expr, envir, enclos): object 'res' not found