Get tidy model output with fast CIs

By default the (amazing) `broom` package uses the `confint()` function to calculate CIs. For GLMs this calculates confidence intervals via profile likelihood by default. When using large datasets this takes a long time and does not meaningfully alter the CIs compared to simply calculating using 1.96*SE. This function `tidy_ci()` runs `broom::tidy()` and returns the tidy estimates with CIs calculated as EST +/- 1.96*SE. (Well, actually 1.959964 :: from `lukesRlib::get_z(0.05)`)

The function also does a few other nice/useful things to the output: hides the intercept by default, automatically detects logistic/CoxPH/CRR models and exponentiates the estimates, and if p=0 returns the extreme p as a string. Other optional outputs include -log10 p-values.

Not tested for models other than `glm()` and `survival::coxph()` where it seems to work very well and produces consistent CIs. Also works well for `cmprsk::crr()`

Usage

tidy_ci(
  x,
  ci = TRUE,
  ci_denominator = 1.959964,
  exp = FALSE,
  intercept = FALSE,
  tidy_factors = TRUE,
  extreme_ps = TRUE,
  neglog10p = FALSE,
  check_model = TRUE,
  n = NA,
  get_r2 = TRUE,
  conf.int = FALSE,
  quiet = FALSE,
  ...
)

Arguments

x

object containing model output to be tidied e.g., from a `glm()` or `survival::coxph()`.

ci

Logical. Default is TRUE. Calculate CIs using 1.96*SE method - where 1.96 can be modified using `ci_denominator`.

ci_denominator

Numeric. Default is 1.96. The standard error of the sample mean (default actually 1.959964 :: from `get_z(0.05)`).

exp

Logical. Default is FALSE. Exponentiate estimate and CIs -- also see `check_model`.

intercept

Logical. Default is FALSE. Exclude intercept for tidier output.

tidy_factors

Logical. Default is TRUE. Tidy `as.factor(x_var)#` terms to `x_var-#`.

extreme_ps

Logical. Default is TRUE. If p=0 then return "extreme p-values" as strings.

neglog10p

Logical. Default is FALSE. Provides negative log10 p-values (if input is class `glm` or `coxph` or `crr` -- user can provide sample size `n=#` to override).

check_model

Logical. Default is TRUE. Set `exp=TRUE` if `glm(family=binomial)` or `survival::coxph()` or `cmprsk::crr()` was performed.

n

Numeric. Default is NA. The N for `neglog10p` is extracted automatically for `glm` or `coxph` objects - override here if required.

get_r2

Logical. Default is TRUE. get R^2 value if model is linear, or C-statistic if model is CoxPH

quiet

Logical. Default is FALSE. Suppress text output

...

Other `tidy()` options

Value

Returns a tibble - summary statistics from a model

Author

Luke Pilling

Examples

fit_linear = glm(bmi ~ age + sex + as.factor(smoking_status), data = d)
#> Error in eval(mf, parent.frame()): object 'd' not found
tidy_ci(fit_linear)
#> Error in eval(expr, envir, enclos): object 'fit_linear' not found

fit_logistic = glm(current_smoker_vs_never ~ age + sex + bmi, data = d, family = binomial(link="logit"))
#> Error in eval(mf, parent.frame()): object 'd' not found
tidy_ci(fit_logistic)   # detect model and exponentiate automatically
#> Error in eval(expr, envir, enclos): object 'fit_logistic' not found
tidy_ci(fit_logistic, check_model=FALSE)  # override auto checking to get untransformed estimates
#> Error in eval(expr, envir, enclos): object 'fit_logistic' not found

fit_coxph = coxph(Surv(time, status) ~ age + sex + as.factor(smoking_status), data = d)
#> Error in coxph(Surv(time, status) ~ age + sex + as.factor(smoking_status),     data = d): could not find function "coxph"
tidy_ci(fit_coxph)
#> Error in eval(expr, envir, enclos): object 'fit_coxph' not found