Get tidy model output with fast CIs

By default the (amazing) `broom` package uses the `confint()` function to calculate CIs. For GLMs this calculates confidence intervals via profile likelihood by default. When using large datasets this takes a long time and does not meaningfully alter the CIs compared to simply calculating using 1.96*SE. This function `tidy_ci()` runs `broom::tidy()` and returns the tidy estimates with CIs calculated as EST +/- 1.96*SE.

The function also does a few other nice/useful things to the output: hides the intercept by default, automatically detects logistic/CoxPH/CRR models and exponentiates the estimates, and if p=0 returns the extreme p as a string. Other optional outputs include -log10 p-values.

Not tested for models other than `glm()` and `survival::coxph()` where it seems to work very well and produces consistent CIs. Also works well for `cmprsk::crr()`

tidy_ci(
  x,
  ci = TRUE,
  exp = FALSE,
  intercept = FALSE,
  extreme_ps = TRUE,
  neglog10p = FALSE,
  check_model = TRUE,
  n = NA,
  conf.int = FALSE,
  ...
)

Arguments

x

object containing model output to be tidied e.g., from a `glm()` or `survival::coxph()`

ci

calculate CIs using 1.96*SE method (default=TRUE)

exp

exponentiate estimate and CIs -- also see `check_model` (default=FALSE)

intercept

Exclude intercept for tidier output (default=FALSE)

extreme_ps

If p=0 then return "extreme p-values" as strings (default=TRUE)

neglog10p

Provides negative log10 p-values (if input is class `glm` or `coxph` or `crr` -- user can provide sample size `n=#` to override) (default=FALSE)

check_model

set `exp=TRUE` if `glm(family=binomial)` or `survival::coxph()` or `cmprsk::crr()` was performed (default=TRUE)

n

the N for `neglog10p` is extracted automatically for `glm` or `coxph` objects - override here if required (default=NA)

...

Other `tidy()` options

Value

Returns a tibble - summary statistics from a model

Author

Luke Pilling

Examples

fit_linear = glm(bmi ~ age + sex + as.factor(smoking_status), data = d)
#> Error in is.data.frame(data): object 'd' not found
tidy_ci(fit_linear)
#> Error in broom::tidy(x, conf.int = conf.int, exponentiate = FALSE, ...): object 'fit_linear' not found

fit_logistic = glm(current_smoker_vs_never ~ age + sex + bmi, data = d, family = binomial(link="logit"))
#> Error in is.data.frame(data): object 'd' not found
tidy_ci(fit_logistic)   # detect model and exponentiate automatically
#> Error in broom::tidy(x, conf.int = conf.int, exponentiate = FALSE, ...): object 'fit_logistic' not found
tidy_ci(fit_logistic, check_model=FALSE)  # override auto checking to get untransformed estimates
#> Error in broom::tidy(x, conf.int = conf.int, exponentiate = FALSE, ...): object 'fit_logistic' not found

fit_coxph = coxph(Surv(time, status) ~ age + sex + as.factor(smoking_status), data = d)
#> Error in coxph(Surv(time, status) ~ age + sex + as.factor(smoking_status),     data = d): could not find function "coxph"
tidy_ci(fit_coxph)
#> Error in broom::tidy(x, conf.int = conf.int, exponentiate = FALSE, ...): object 'fit_coxph' not found